move simple base models into models.py
