+++ src/myteam/workflow/models.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import Any, Literal, Optional, TypedDict
 
-from pydantic import BaseModel, ConfigDict, Field, PositiveInt
+from pydantic import BaseModel, ConfigDict, Field, PositiveInt, RootModel, model_validator
 
 
 class AgentConfig(TypedDict):
@@ -47,6 +47,59 @@ class ProjectWorkflowDefaults(BaseModel):
     extra_args: Optional[list[str]] = Field(default=None)
     usage_logging: Optional[Literal["none", "summary", "per_model", "verbose"]] = Field(default=None)
     inactivity_timeout_seconds: Optional[PositiveInt] = Field(default=None)
+
+
+class StepDefinitionModel(BaseModel):
+    model_config = ConfigDict(extra="forbid", strict=True)
+
+    prompt: str = Field(min_length=1)
+    output: dict[str, Any] = Field(default_factory=dict)
+    input: Any = None
+    agent: Optional[str] = Field(default=None, min_length=1)
+    model: Optional[str] = Field(default=None, min_length=1)
+    extra_args: list[str] | None = None
+    interactive: bool | None = None
+    session_id: Optional[str] = Field(default=None, min_length=1)
+    fork: bool | None = None
+
+    @model_validator(mode="after")
+    def _fork_requires_session(self) -> "StepDefinitionModel":
+        if self.fork and self.session_id is None:
+            raise ValueError("field 'session_id' is required when 'fork' is true.")
+        return self
+
+    @model_validator(mode="after")
+    def _agent_must_exist(self) -> "StepDefinitionModel":
+        if self.agent is not None:
+            from .agents import get_agent_config
+
+            try:
+                get_agent_config(self.agent)
+            except KeyError as exc:
+                raise ValueError(str(exc)) from exc
+        return self
+
+
+class WorkflowDefinitionModel(RootModel[dict[str, StepDefinitionModel]]):
+    model_config = ConfigDict(strict=True)
+
+
+class StepExecutionArgs(BaseModel):
+    model_config = ConfigDict(extra="forbid", strict=True)
+
+    input: Any = None
+    agent_name: str = Field(min_length=1)
+    interactive: bool = True
+    session_id: Optional[str] = Field(default=None, min_length=1)
+    fork: bool = False
+    extra_args: list[str] | None = None
+    model: Optional[str] = Field(default=None, min_length=1)
+
+    @model_validator(mode="after")
+    def validate_fork_requires_session(self) -> "StepExecutionArgs":
+        if self.fork and self.session_id is None:
+            raise ValueError("session_id is required when fork is true")
+        return self
 
 
 @dataclass
